+++ docs/changelog.md
@@ -1,9 +1,13 @@
 # TBD
-###Features
+### Features
 * Created the Kurtosis project boilerplate which includes: CI configuration, build and run script and release script
 * Added testsuite execution configuration with a custom configurator `my_testsuite_configurator` and a custom testsuite `my_testsuite`
 * Added instructions for implementing the basic ETH test into the `README.md` file which test GETH client execution following the GETH documentation for [setting up a GETH client in DEV mode](https://geth.ethereum.org/docs/getting-started/dev-mode)
 * Added instructions for implementing the advanced ETH test into the `README.md` file which test an Ethereum private network following the GETH documentation for [setting up an Ethereum private network](https://geth.ethereum.org/docs/interface/private-network)
 
-#0.1.0
+### Changes
+* Simplified the basic test instructions and fixed some bugs
+* Renamed `MyTest` to `BasicEthereumTest`
+
+# 0.1.0
 * Init commit